Revert "Adding docker and gulp to the project (#19)"
This reverts commit c8e0788fe5c5ebdc45c9b7ceded064bb8f9161b3.
