--- .eslintrc.js
@@ -1,24 +1,0 @@
-module.exports = {
-    "env": {
-        "browser": true
-    },
-
-    "rules": {
-        "indent": [
-            "error",
-            "tab"
-        ],
-        "linebreak-style": [
-            "error",
-            "unix"
-        ],
-        "quotes": [
-            "error",
-            "single"
-        ],
-        "semi": [
-            "error",
-            "always"
-        ]
-    }
-};
