+++ .gitignore
@@ -64,8 +64,4 @@ $RECYCLE.BIN/
 # Windows shortcuts
 *.lnk
 
-node_modules/
-build/
 # End of https://www.gitignore.io/api/linux,macos,windows
-
-.idea/
