--- .travis.yml
@@ -1,11 +1,0 @@
-sudo: required
-
-language: node_js
-node_js:
-  - "0.10"
-
-install:
-  - npm install -g eslint
-
-script:
-  - npm run lint
