+++ README.md
@@ -1,19 +1,17 @@
 # <img src="https://github.com/Addy-Smart-Addresses/Addy/blob/master/img/logo.png" alt="Addy" height="42" width="42"></img>   Addy
-Addy is a web app that enhances the addressing system of the world. Instead of remembering long addresses, Addy generates a unique 7-alphanumeric code based on the latitudinal and longitudinal coordinates of any position. The code is so accurate that every building in the world has been assigned a different 'Addy' with a minimum distance of 11.132 meters.
+Addy is a web app that enhances the addressing system of the world. Instead of remembering long addresses, Addy generates a unique 7-digit code based on the latitudinal and longitudinal coordinates of any position. The code is so accurate that every building in the world has been assigned a different 'Addy' with a minimum distance of 11.132 meters.
 
 ![Landing Page](readme_assets/home.png)
 
-
-**Tech Stack for this project**
-- HTML5
-- CSS3
-- JavaScript
-- Docker
-- docker-compose
-- Gulp
-- BootStrap 3
-- Google Maps Places API
-- Google Maps Geocoder API Used
+<strong>Tech Stack for this project</strong>
+<ul>
+  <li>HTML5</li>
+  <li>CSS3</li>
+  <li>JavaScript</li>
+  <li>BootStrap 3</li>
+  <li>Google Maps Places API</li>
+  <li>Google Maps Geocoder API Used</li>
+</ul>
 
 ## Gitter channel link
 https://gitter.im/Addy-Smart-Addresses
@@ -46,11 +44,3 @@ git commit -m <insert_commit_comment>
 git push -u origin <your_branch_name>
 ```
 8. Go to the GitHub repository and submit a Pull Request.
-
-## Running
-
-Once you have [docker](https://docs.docker.com/engine/installation/) and [docker-compose](https://docs.docker.com/compose/install/) installed:
-
-    docker-compose up --build
-    
-A development version will be running at port 8000 and a simulation of production build at port 8001.
