--- docker-compose.yml
@@ -1,18 +1,0 @@
-version: '3'
-
-services:
-  addy-development:
-    build: .
-    command: /bin/bash -l -c "python -m SimpleHTTPServer 8000"
-    tty: true
-    ports:
-      - 8000:8000
-    volumes:
-      - $PWD/src:/usr/src
-      - /usr/src/node_modules
-  addy-production:
-    build: .
-    command: /bin/bash -l -c "gulp && cd build/ && python -m SimpleHTTPServer 8001"
-    tty: true
-    ports:
-      - 8001:8001
