+++ index.html
@@ -1,0 +1,75 @@
+<html>
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <meta http-equiv="x-ua-compatible" content="ie=edge">
+  <link rel="shortcut icon" sizes="16x16" href="img/logo-16.ico">
+  <link rel="shortcut icon" sizes="32x32" href="img/logo-32.ico">
+  <title>Addy</title>
+<link rel="stylesheet" href="maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/css/bootstrap.min.css" integrity="sha384-/Y6pD6FV/Vv2HJnA6t+vslU6fwYXjCFtcEpHbNJ0lyAFsXTsjBbfaDjzALeQsN6M"
+  crossorigin="anonymous">
+<script type="text/javascript" src="maps.google.com/maps/api/js?key=AIzaSyCEqoC5yQIIBXbNVxsuRe787A29ZC32CVA&libraries=places"></script>
+
+
+
+
+  <link href="maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
+  <link href="fonts.googleapis.com/css?family=Muli" rel="stylesheet">
+  <link type="text/css" href="./css/style.css" rel="stylesheet">
+</head>
+
+<body onload="codeAddress()">
+  <div id="maparea" align="center"></div>
+
+  <img src="img/logo.png" class="img-responsive logo" onclick="getLocation()" />
+  <div class="sidebar">
+	<div class="input-group">
+	  <input class="form-control" id="NameSearch" type="textbox" value="" onkeydown="if (event.keyCode == 13){codeAddress()}"
+		placeholder="Search by Name">
+	  <span class="input-group-addon" onclick="codeAddress()"><i class="fa fa-search"></i></span>
+	</div>
+
+	  <center class="text-muted my-2">OR</center>
+	  <div class="input-group">
+		<input class="form-control input-md" id="BackCode" type="textbox" placeholder="Search By Addy" onclick="document.getElementById('BackCode').value=''"
+		  onkeydown="if (event.keyCode == 13){codeback()}"></input>
+		<span class="input-group-addon" onclick="codeback()"><i class="fa fa-search"></i></span>
+	  </div>
+
+	  <hr>
+
+	  <div class="info">
+		<div class="bar">
+		  <div class="content" id="addycode">4J3423YX</div>
+		  <div class="heading">Addy Code</div>
+		</div>
+		<div class="bar">
+		  <div class="content" id="address">Unnamed Road, Sector 127, Noida, Uttar Pradesh 201303, India</div>
+		  <div class="heading">Address</div>
+		</div>
+		<div class="bar">
+		  <div class="content" id="info">28.5359744, 77.34566410000002</div>
+		  <div class="heading">Info</div>
+		</div>
+	  </div>
+
+	  <center>
+		<button type="button" class="btn btn-primary mt-5 btn-block " id="directions">Get Directions from your Location</button>
+	  </center>
+
+  </div>
+
+
+
+
+
+<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
+  crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js" integrity="sha384-b/U6ypiBEHpOf/4+1nzFpr53nxSS+GLCkfwBdFNTxtclqqenISfwAzpKaMNFNmj4"
+  crossorigin="anonymous"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/js/bootstrap.min.js" integrity="sha384-h0AbiXch4ZDo7tp9hKZ4TsHbi047NrKGLO3SEJAg45jXxnGIfYzk4Si90RDIqNm1"
+  crossorigin="anonymous"></script>
+
+<script type="text/javascript" src="./js/script.js"></script>
+</body>
+</html>
