--- package.json
@@ -1,33 +1,0 @@
-{
-  "name": "addy",
-  "version": "1.0.0",
-  "description": "Addy is a web app that enhances the addressing system of the world. Instead of remembering long addresses, Addy generates a unique 7-digit code based on the latitudinal and longitudinal coordinates of any position. The code is so accurate that every building in the world has been assigned a different 'Addy' with a minimum distance of 11.132 meters.",
-  "main": "index.js",
-  "scripts": {
-    "lint": "eslint /src",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Addy-Smart-Addresses/Addy.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Addy-Smart-Addresses/Addy/issues"
-  },
-  "homepage": "https://github.com/Addy-Smart-Addresses/Addy#readme",
-  "devDependencies": {
-    "gulp": "^3.9.1",
-    "gulp-csso": "^3.0.0",
-    "gulp-htmlmin": "^3.0.0",
-    "gulp-imagemin": "^3.4.0",
-    "gulp-size": "^2.1.0",
-    "gulp-uglify": "^3.0.0",
-    "require-dir": "^0.3.2"
-  },
-  "dependencies": {
-    "eslint": "^4.11.0",
-    "gulp-webpack": "^1.5.0"
-  }
-}
