--- src/gulp/tasks/default.js
@@ -1,3 +1,0 @@
-var gulp = require('gulp');
-
-gulp.task('default', ['optimize:css', 'optimize:html', 'optimize:js', 'optimize:images']);
