--- src/gulp/tasks/production/optimize-html.js
@@ -1,12 +1,0 @@
-var gulp    = require('gulp');
-var htmlmin = require('gulp-htmlmin');
-var config  = require('../../config').optimize.html;
-
-/**
- * Minimize HTML
- */
-gulp.task('optimize:html', function() {
-	return gulp.src(config.src)
-		.pipe(htmlmin(config.options))
-		.pipe(gulp.dest(config.dest));
-});
