--- src/gulp/tasks/production/optimize-js.js
@@ -1,20 +1,0 @@
-var gulp   = require('gulp');
-var uglify = require('gulp-uglify');
-var size   = require('gulp-size');
-var webpack= require('gulp-webpack');
-var config = require('../../config').optimize.js;
-
-/**
- * Copy and minimize JS files
- */
-gulp.task('optimize:js', function() {
-  return gulp.src(config.src)
-  	.pipe(webpack({
-  		output: {
-  			filename: "script.js"
-  		}
-  	}))
-    .pipe(uglify(config.options))
-    .pipe(gulp.dest(config.dest))
-    .pipe(size());
-});
