--- src/gulpfile.js
@@ -1,4 +1,0 @@
-var requireDir = require('require-dir');
-
-// Require all tasks in gulp/tasks, including subfolders
-requireDir('./gulp/tasks', { recurse: true });
